Improved damage calculations. Extended and improved API a LOT.

+++ src/main/java/com/mineshaft/mineshaftapi/manager/player/PlayerStatManager.java
@@ -109,7 +109,7 @@ public class PlayerStatManager {
     }
 
     public static int getProficiencyBonus(int level) {
-        return (int) (1+ (level-1)*4);
+        return 1+(level-1)*4;
     }
 
 }
